refactor: rename genAIIndex to genai
