+++ src/namespaces/genai/genai.ts
